feat: Spawn CKB node directly instead of running it from a background window renderer process
