--- packages/neuron-wallet/src/startup/ckb-runner/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title></title>
-</head>
-<body>
-  <script>
-    require('./task.js')
-  </script>
-</body>
-</html>
